requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
pydantic
